Dashboard/requirements.txt
click==8.0.1
Flask==2.0.1
hurry.filesize==0.9
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pymongo==3.11.4
python-dotenv==0.18.0
Werkzeug==2.0.1
flask_socketio==5.3.4
gevent-websocket==0.10.1
flask_cors==3.0.10